deluxe_macros

Derive Macro ParseAttributes

#[derive(ParseAttributes)]
{
    // Attributes available to this derive:
    #[deluxe]
}

Generates ParseAttributes for a struct or enum.

Supported attributes are the same as in ParseMetaItem. Additionally, some extra attributes are supported.

Extra Container Attributes

The following extra attributes are supported on structs and enums:

#[deluxe(attributes(...))]

Provides a list of paths to match attribute names against. These paths will be checked inside the ParseAttributes::path_matches function to see if a path matches. This attribute can be specified multiple times to append more paths to the list of matching paths.

Omitting this attribute will cause all paths to match. This can be useful when doing your own filtering on an attribute list before parsing.

Extra Field Attributes

The following extra attributes are supported on struct fields and enum fields:

#[deluxe(container)]

Specifies that this field should store the current obj being parsed by parse_attributes. The field can be a value type, an Option, or a reference type. If the field is not a reference type, the container will be cloned into the field.

If the type of the container contains a lifetime, Deluxe will try to use it for a lifetime bound for ParseAttributes. If the container type is a reference, it will try to infer it from the lietime of the reference. Otherwise, it will try to take the first lifetime parameter used by the type if one is present. If no lifetimes can be inferred, a new lifetime parameter will be generated. See container(lifetime = ...) for instructions on how to specify a lifetime parameter manually.

If the struct/enum also derives ParseMetaItem, then #[deluxe(default)] is implied on any container fields. In that case, a common pattern is to use an Option as the container field. It will be set to None when calling parse_meta_item, but will be Some when calling parse_attributes.

If used within an enum, only the first container field will supply the trait bounds. Any other container fields in other variants must have a compatible type and lifetime.

Fields with this attribute can safely be added to a struct or variant using #[deluxe(transparent)]. Container fields do not count as a parseable field, as they are never parsed from tokens.

#[deluxe(container(lifetime = 't, ty = path::MyType)]

Specifies that this field should store the current obj being parsed by parse_attributes, with a custom lifetime and type bound used for the ParseAttributes trait implementation.

Normally the lifetime and ty are not needed when using #[deluxe(container)] because the macro can infer the lifetime and type from the field itself. If Deluxe is unable to infer them, these attributes can be supplied to manually provide the lifetime and type. lifetime can be omitted if the container is an owning type.

This attribute is implemented by calling methods on the ContainerFrom trait. Other container types besides references and Option will also need to provide an implementation for that trait.