deluxe_macros
Derive Macro ExtractAttributes

#[derive(ExtractAttributes)]
{
    // Attributes available to this derive:
    #[deluxe]
}

Generates ExtractAttributes for a struct or enum.

This macro is identical to ParseAttributes, except for the following differences:

The generated extract_attributes function will remove any attributes matching the paths listed in #[deluxe(attributes(...))].

Reference types for #[deluxe(container)] fields will not work. The container type must be an owned type or an Option, and the container will always be cloned into the field. The cloning will happen after the attributes have been extracted. Using the lifetime parameter for custom container types will also not work.

See the ParseAttributes and ParseMetaItem documentation for all other details.